new update

+++ back-end/controller/userController.js
@@ -11,10 +11,10 @@ exports.registerUser = catchAsyncError(async (req, res, next) => {
   let avatar;
   if (JSON.stringify(req.files) === JSON.stringify({})) {
     return next(new ErrorHandler("Please Upload Image", 400));
-  } else {
-    console.log(req.files);
+  } 
+    // const  imageName = await upload(req.files.avatar[0]);
     avatar = req.files.avatar[0].path;
-  }
+ 
   const { name, email, password } = req.body;
   const user = await User.create({
     name,
@@ -75,7 +75,9 @@ exports.forgotPassword = catchAsyncError(async (req, res, next) => {
 
   await user.save({ validateBeforeSave: false });
   //                                          req.get("host")
-  const resetPasswordUrl = `${req.protocol}://${req.get("host")}/password/reset/${resetToken}`;
+  const resetPasswordUrl = `${req.protocol}://${req.get(
+    "host"
+  )}/password/reset/${resetToken}`;
 
   const message = `<p>Your password reset token is :- \n\n${resetPasswordUrl}<br />If you have not requested then please ignore it</p>`;
 
@@ -136,6 +138,10 @@ exports.resetPassword = catchAsyncError(async (req, res, next) => {
 exports.getUserDetails = catchAsyncError(async (req, res, next) => {
   const user = await User.findById(req.user.id);
 
+  // const url = await download(user.avatar);
+
+  // user.avatar = url;
+
   return res.status(200).json({
     success: true,
     user,
@@ -164,12 +170,16 @@ exports.updatePassword = catchAsyncError(async (req, res, next) => {
 });
 
 exports.updateProfile = catchAsyncError(async (req, res, next) => {
-  const oldUser = await User.findById(req.user.id);
+  const user = await User.findById(req.user.id);
 
   let newUserData;
 
   if (JSON.stringify(req.files) !== JSON.stringify({})) {
-    fs.unlinkSync(oldUser.avatar);
+  //  await deleteImage(user.avatar);
+  //  const  imageName  = await upload(req.files.avatar[0]);
+  if (fs.existsSync(user.avatar)) {
+    fs.unlinkSync(user.avatar);
+  }
     newUserData = {
       name: req.body.name,
       email: req.body.email,
@@ -182,7 +192,7 @@ exports.updateProfile = catchAsyncError(async (req, res, next) => {
     };
   }
 
-  const user = await User.findByIdAndUpdate(req.user.id, newUserData, {
+  await User.findByIdAndUpdate(req.user.id, newUserData, {
     new: true,
     runValidators: true,
     useFindAndModify: false,
@@ -249,8 +259,10 @@ exports.deleteUser = catchAsyncError(async (req, res, next) => {
     );
   }
 
-  fs.unlinkSync(user.avatar);
-
+  // await deleteImage(user.avatar);
+  if (fs.existsSync(user.avatar)) {
+    fs.unlinkSync(user.avatar);
+  }
   await user.remove();
 
   return res.status(200).json({
